Only retrieve the top 10 items in the build queue at a time.
Previously we'd retrieve every candidate BuildQueue.id in order and 
invoke postprocessCandidate on each until we found one that passed. But
in the real world there's always a valid candidate near the top of the 
queue, postprocessCandidate unqueues any jobs that it rejects, and we 
want bounded per-cycle query times anyway. If we do end up with 1000 
unusable jobs at the top of the queue, we just have to wait a minute or
so for them to be killed, 10 at a time.