Only set the build status at the same time buildd-manager dequeues a build.
BuildFarmJobBehaviour.handleSuccess previously called build.updateStatus
and committed itself, meaning that the status was set for a (usually)
short period before the BuildQueue was destroyed and buildd-manager
stopped considering the build. process-upload considers status=UPLOADING
to constitute handoff, so the two could operate on the same build
concurrently. Instead, handleSuccess now returns a status for
handleStatus to set atomically as it dequeues the build.