Split UEFI unit-tests into command validation and functional tests
Firstly we test that the various core functions (genUefiKeys, signUefi)
will generate appropriately formed commands based on their inputs.

Secondly we test that an appropriate set of core calls are made based
on the input forms we generate.  This way we avoid having to understand
the commands generated in order to validate functionality.

Signed-off-by: Andy Whitcroft <apw@ubuntu.com>