unit-tests: switch FakeMethodGenUefiKeys to use upload paths
We are checking that the keys get built.  We check those against our expected
paths, but we should be faking the build of those files against the paths
the upload object would have used.  This allows us to confirm it is putting
them in the expected paths later.  Switch to passing in the upload object so
we can see its paths.

Signed-off-by: Andy Whitcroft <apw@ubuntu.com>