Don't hash duplicates in librarian-gc.
Instead of comparing everything, just check that the original file's
hash matches the DB and trust that the rest haven't collided.
Pedantically verifying that every duplicate file matches the original
file bit by bit is very slow now that we have a lot of duplicates and
most of them have to be downloaded from Swift.