Use a per-archive OOPS timeline in archivepublisher scripts.
Previously SQL queries weren't logged, just librarian and co. requests,
and the timeline covered the entire script run. This resulted in
OOPSes with timelines that were 99.99% irrelevant and omitted the
important SQL bits.

SQL queries are now logged, and the timeline is reset between archives
in process-accepted, publish-distro and process-death-row.