Fix builds_for_snap to avoid iterating over an unsliced DecoratedResultSet.
builds_for_snap iterates over snap.pending_builds and snap.completed_builds
doing visibility checks, and terminates the loop over snap.completed_builds
if it reaches ten builds.  Unfortunately, the first call to
snap.completed_builds.__iter__ runs the pre_iter_hook on the whole
DecoratedResultSet, since we aren't slicing it, which does a great deal of
work if the snap has many builds.

As it happens, the model already does the necessary visibility checks, as
proven by the existing tests for this still passing.  So we can just drop
this excess complication, slice snap.completed_builds as necessary, and be
done.