Precache permissions for archives returned by Person.getVisiblePPAs.
This fixes large query counts on Person:+index in cases where the requesting
user has upload permission on many PPAs owned by the context person.

I made the same change to Person:+activate-ppa for completeness, although
that can't currently cause a problem in practice because it can only be used
by the context person or by admins, thereby bypassing most of the slow bits
of the security adapter.