Speed up tests that use archive signing keys using an in-process keyserver fixture.
This only works for tests that only talk to the keyserver asynchronously,
but for those that do it's very much faster to start one up in-process
rather than using a .tac file: it saves on the order of 10 seconds per
affected test on my laptop.

The tests for the fixture itself seem to hit some slightly buggy bit of
Twisted and require a couple of extra spins through the reactor to clear up
cancelled DelayedCalls, but fortunately testtools.deferredruntest already
has machinery to tolerate that.