Fix git-to-git code imports on xenial.
Deleting the remote-tracking refs after the fetch now requires
'refs/remotes/source/**' rather than 'refs/remotes/source/*' following
changes to git's wildcard-matching implementation.

As a result of this bug having been live on production for a few days, we
also need to delete any such refs that may exist before fetching, as the
conflict between refs on disk and refs in the remote seems to confuse git in
some cases.