Avoid importing testtools.content outside tests.
lp_sitecustomize imports lp.services.log.logger, so importing testtools
there is expensive.  Moving BufferLogger to a test-only location turns out
to be somewhat difficult as a couple of non-test modules currently use it,
but we can at least only import testtools.content in the one method that
needs it.