Special-case auto_build_channels={"snapcraft": "apt"} to install snapcraft using apt.
Without something like this, it's hard to do a well-managed migration to
installing snapcraft from the stable channel by default.  We're going to
need a way for individual snaps to opt out in case they don't work with
snapcraft as a snap for whatever reason, and if the unmarked default case
(i.e.  `auto_build_channels=None`) becomes equivalent to `{"snapcraft":
"stable"}` then they need some other escape route.

The alternative would be to do a data migration to set `{"snapcraft":
"stable"}` for everything, but that's hard to revert if something goes wrong
and isn't in the spirit of feature flags.