Allow macaroon authentication to the librarian for BPBs' source files.
This will later allow us to dispatch private builds immediately, rather than
having to wait for the source to be published first.

In this case we accept the macaroon as the password part of basic
authentication with an empty username, similar to the git authentication
method used by code import jobs.  This is a bit weird, but it avoids needing
to change launchpad-buildd, and we could always switch to a more
conventional "Authorization: Macaroon ..." scheme later.