Convert the Roundup external bug tracker to urlfetch.
I experimented with a couple of different approaches to converting the
doctests, and eventually decided that the most readable option was to keep
TestRoundup and add some responses-based helper code to it.  This is done in
such a way that it can be reused for the other test tracker subclasses.