Refactor traversal helpers to avoid using Zope class advice.
Zope class advice doesn't work in Python 3, and decorators seem somewhat
more Pythonic anyway.  We just need to store the information in method
attributes rather than class attributes.

I considered retaining the two-argument form of redirection as a class
decorator, but rewriting the call sites was trivial enough that it wasn't
worth it.