Add GitPermissionType and GitRuleGrant.permissions abstractions.
At the moment, the new enumeration is mainly just a minor convenience for
the GitAPI XML-RPC endpoint, which can now work with an enumeration
internally rather than ad-hoc strings and only serialise them when producing
a method response.  However, the UI will need to have a vocabulary of common
combinations of permissions, and the enumeration will make that much easier.

Similarly, GitRuleGrant.permissions currently just makes it slightly easier
for GitAPI to extract all the permissions in one go, but it will make it
much easier for the UI to work with permissions.