Improve close-account's foreign key handling.
I went through the list of FKs to Person.id and identified quite a few more
that can safely be either ignored or removed.  For the remainder, I borrowed
part of personmerge's approach: we now programmatically check all remaining
FKs and bail out if there's anything left that we haven't considered.  (For
example, if the person whose account is being closed still owns a project,
we need to decide what to do about that.)