Add logging to macaroon verification, and refactor tests to check it.
It was all too easy for macaroon verification to fail for some reason other
than the one that a test expected, and tests had no way to tell.  We now log
details of verification failures, and I added a couple of test helpers to
make it easy for tests to check these consistently.