Refactor code import authorisation.
We now use the special "+launchpad-services" requester ID for code import
job pushes, which makes somewhat more sense than treating them as a special
kind of anonymous request.

If macaroon authentication fails, we now explicitly return an Unauthorized
fault rather than falling through to end up with either PermissionDenied for
a public repository or GitRepositoryNotFound for a public one.  This is
easier to understand and also more consistent with RFC 2616's description of
"401 Unauthorized".